Refactor dialog to its own component
