+++ components/d2l-activity-editor/d2l-activity-score-editor.js
@@ -1,4 +1,5 @@
 import './d2l-activity-grades/d2l-activity-grade-candidate-selector';
+import './d2l-activity-grades-dialog.js';
 import '@brightspace-ui/core/components/button/button.js';
 import '@brightspace-ui/core/components/colors/colors.js';
 import '@brightspace-ui/core/components/dialog/dialog.js';
@@ -8,16 +9,13 @@ import '@brightspace-ui/core/components/inputs/input-text.js';
 import 'd2l-dropdown/d2l-dropdown.js';
 import 'd2l-dropdown/d2l-dropdown-menu.js';
 import 'd2l-tooltip/d2l-tooltip';
-import { bodyCompactStyles, bodySmallStyles, labelStyles } from '@brightspace-ui/core/components/typography/styles.js';
+import { bodyCompactStyles, labelStyles } from '@brightspace-ui/core/components/typography/styles.js';
 import { css, html } from 'lit-element/lit-element';
 import { ActivityEditorMixin } from './mixins/d2l-activity-editor-mixin.js';
-import { formatNumber } from '@brightspace-ui/intl/lib/number.js';
 import { getLocalizeResources } from './localization';
-import { shared as gradeCandidateCollectionStore } from './d2l-activity-grades/state/grade-candidate-collection-store.js';
 import { inputStyles } from '@brightspace-ui/core/components/inputs/input-styles.js';
 import { LocalizeMixin } from '@brightspace-ui/core/mixins/localize-mixin.js';
 import { MobxLitElement } from '@adobe/lit-mobx';
-import { radioStyles } from '@brightspace-ui/core/components/inputs/input-radio-styles.js';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
 import { shared as store } from './state/activity-store.js';
 
@@ -33,10 +31,8 @@ class ActivityScoreEditor extends ActivityEditorMixin(LocalizeMixin(RtlMixin(Mob
 	static get styles() {
 		return [
 			bodyCompactStyles,
-			bodySmallStyles,
 			labelStyles,
 			inputStyles,
-			radioStyles,
 			css`
 			:host {
 				display: block;
@@ -141,20 +137,6 @@ class ActivityScoreEditor extends ActivityEditorMixin(LocalizeMixin(RtlMixin(Mob
 				font-family: inherit;
 				color: inherit;
 			}
-			.grade-dialog-create-new-container {
-				display: flex;
-			}
-			.grade-dialog-create-new-activity-name {
-				word-break: break-word;
-				overflow-wrap: anywhere; /* not supported by safari */
-			}
-			.grade-dialog-create-new-icon {
-				padding-right: 10px;
-			}
-			:host([dir="rtl"]) .grade-dialog-create-new-icon {
-				padding-right: 0;
-				padding-left: 10px;
-			}
 			`
 		];
 	}
@@ -228,26 +210,10 @@ class ActivityScoreEditor extends ActivityEditorMixin(LocalizeMixin(RtlMixin(Mob
 	}
 
 	async _chooseFromGrades() {
-		const scoreAndGrade = store.get(this.href).scoreAndGrade;
-		const existingAssociatedGradeHref = scoreAndGrade.gradeHref;
-		const dialog = this.shadowRoot.querySelector('d2l-dialog');
-
-		const action = await dialog.open();
-		if (action !== 'done') {
-			return;
+		const activityGradesElement = this.shadowRoot.querySelector('d2l-activity-grades-dialog');
+		if (activityGradesElement) {
+			await activityGradesElement.open();
 		}
-
-		this._associateToExistingGrade(scoreAndGrade, existingAssociatedGradeHref);
-	}
-
-	_associateToExistingGrade(scoreAndGrade, existingAssociatedGradeHref) {
-		const gradeCandidateCollection = gradeCandidateCollectionStore.get(scoreAndGrade.gradeCandidatesHref);
-		const newGradeCandidateSelected = gradeCandidateCollection.selected;
-		if (existingAssociatedGradeHref && existingAssociatedGradeHref === newGradeCandidateSelected.href) {
-			return;
-		}
-
-		scoreAndGrade.setAssociatedGrade(newGradeCandidateSelected);
 	}
 
 	render() {
@@ -263,8 +229,7 @@ class ActivityScoreEditor extends ActivityEditorMixin(LocalizeMixin(RtlMixin(Mob
 			gradeType,
 			inGrades,
 			isUngraded,
-			canSeeGrades,
-			gradeCandidatesHref
+			canSeeGrades
 		} = activity.scoreAndGrade;
 
 		this._focusUngraded = isUngraded;
@@ -327,40 +292,10 @@ class ActivityScoreEditor extends ActivityEditorMixin(LocalizeMixin(RtlMixin(Mob
 								</d2l-menu>
 							</d2l-dropdown-menu>
 						</d2l-dropdown>
-						<d2l-dialog title-text="${this.localize('chooseFromGrades')}">
-							<label class="d2l-input-radio-label">
-								<input
-									type="radio"
-									name="chooseFromGrades">
-								${this.localize('createAndLinkToNewGradeItem')}
-							</label>
-							<d2l-input-radio-spacer>
-								<div class="grade-dialog-create-new-container">
-									<div class="grade-dialog-create-new-icon"><d2l-icon icon="tier1:grade"></d2l-icon></div>
-									<div>
-										<div class="grade-dialog-create-new-activity-name">${this.activityName}</div>
-										<div class="d2l-body-small">${scoreOutOf && !scoreOutOfError ? html`
-											${this.localize('points', { points: formatNumber(scoreOutOf, { maximumFractionDigits: 2 })})}
-										` : null }
-										</div>
-									</div>
-								</div>
-							</d2l-input-radio-spacer>
-							<label class="d2l-input-radio-label">
-								<input
-									type="radio"
-									name="chooseFromGrades">
-								${this.localize('linkToExistingGradeItem')}
-							</label>
-							<d2l-input-radio-spacer>
-								<d2l-activity-grade-candidate-selector
-									href="${gradeCandidatesHref}"
-									.token="${this.token}">
-								</d2l-activity-grade-candidate-selector>
-							</d2l-input-radio-spacer>
-							<d2l-button slot="footer" primary dialog-action="done">${this.localize('ok')}</d2l-button>
-							<d2l-button slot="footer" dialog-action="cancel">${this.localize('cancel')}</d2l-button>
-						</d2l-dialog>
+						<d2l-activity-grades-dialog
+							href="${this.href}"
+							.token="${this.token}"
+							.activityName="${this.activityName}"></d2l-activity-grades-dialog>
 					</div>
 				` : null}
 			</div>
